# [IMP] micro-optimizations on model