@robodoo 2025-02-07
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/196944.png)](https://mergebot.odoo.com/odoo/odoo/pull/196944)